+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: dpalmese <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/19 12:05:09 by dpalmese          #+#    #+#             */
-/*   Updated: 2024/02/03 11:27:12 by rizz             ###   ########.fr       */
+/*   Updated: 2024/02/03 15:27:53 by rizz             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #ifndef LIBFT_H
@@ -78,5 +78,5 @@ char				*ft_strtrim(char const *s1, char const *set);
 void				ft_putnbr_fd(int n, int fd);
 char				**ft_split(char const *s, char c);
 size_t				ft_putnstr_fd(char *s, int len, int fd);
-int					ft_putnbrbase_fd(int number, int base, int fd);
+int					ft_putnbrbase_fd(unsigned long number, int base, int fd);
 #endif
